[C++] Fixes for various C++ warnings, added support for warnings as errors in build script

+++ aeron-client/src/main/cpp/ControlledFragmentAssembler.h
@@ -96,7 +96,7 @@ private:
         {
             if ((flags & FrameDescriptor::BEGIN_FRAG) == FrameDescriptor::BEGIN_FRAG)
             {
-                auto result = m_builderBySessionIdMap.emplace(header.sessionId(), m_initialBufferLength);
+                auto result = m_builderBySessionIdMap.emplace(header.sessionId(), static_cast<std::uint32_t>(m_initialBufferLength));
                 BufferBuilder& builder = result.first->second;
 
                 builder
